+++ data/src/main/kotlin/com/vultisig/wallet/data/api/models/GraphQLReponse.kt
@@ -1,0 +1,14 @@
+package com.vultisig.wallet.data.api.models
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class GraphQLResponse<T>(
+    val data: T? = null,
+    val errors: List<GraphQLError>? = null
+)
+
+@Serializable
+data class GraphQLError(
+    val message: String,
+)
